requirements.txt
Django==4.0.3
psycopg2==2.9.3
channels==3.0.4
channels-redis==3.4.0
redis==4.2.1
djangorestframework==3.13.1